refactor: deck 을 통해서 최초 카드를 draw 하도록 변경
